+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Move to Dagger assisted inject
 - Make instant search by name case insensitive
 - Change `InstantSearchScreen` to a fragment
+- Change `ScanSimpleIdScreen` to a fragment
 - [In Progress: 20 Jan 2021] Material Theming Migration
 - [In Progress: 23 Dec 2020] Migrate app to use ViewBinding
 - [In Progress: 13 Jan 2020] Migrate `TheActivity` to the new navigation framework